fix:task-011:グラフ画面の空白調整
