--- lib/features/graph/widget/details_listview.dart
@@ -1,36 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:kakeibo/core/utils/get_category.dart';
-import 'package:kakeibo/features/graph/utils/get_pie_chart_color.dart';
-import 'package:kakeibo/features/graph/utils/get_proportion.dart';
-import 'package:kakeibo/features/graph/widget/category_card.dart';
-
-class DetailsListview extends StatelessWidget {
-  final Map<int, double> categoryDetailsMap;
-  final int total;
-
-  const DetailsListview({
-    super.key,
-    required this.categoryDetailsMap,
-    required this.total,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 600,
-      child: ListView(
-        physics: NeverScrollableScrollPhysics(),
-        children: categoryDetailsMap.entries.map((entry) {
-          final category = entry.key;
-          final value = entry.value;
-          return CategoryCard(
-            value: value.round(),
-            title: getCategory(2, category),
-            color: getPieChartColor(category),
-            proportion: getProportion(value, total),
-          );
-        }).toList(),
-      ),
-    );
-  }
-}
